requirement.txt
langgraph
langsmith
fastapi
langchain
openai 
python-docx 
PyPDF2 
fpdf 
python-dotenv


